Add Combat and game menu

+++ src/survival/main/ui/GUI.java
@@ -12,6 +12,7 @@ import java.awt.*;
 
 import survival.main.Main;
 import survival.main.entity.creatures.Player;
+import survival.main.utils.Util;
 
 /**
  * File: GUI.java 
@@ -27,14 +28,14 @@ public class GUI {
 	
 	private Player player;
 	private int healthBarWidth = 800;
-	private int healthToWidth = healthBarWidth;
+	private double healthToWidth = healthBarWidth;
 	
 	public GUI(Player player) {
 		this.player = player;
 	}
 
 	public void tick() {
-		this.healthToWidth = (int) (player.getHealth() * .01) * healthBarWidth;
+		this.healthToWidth = (player.getHealth() * .01) * healthBarWidth;
 	}
 	
 	public void render(Graphics2D g) {
@@ -42,7 +43,7 @@ public class GUI {
 		g.setColor(Color.black);
 		g.fillRect(((Main.WIDTH / 2) - healthBarWidth / 2) - 2, Main.HEIGHT - 102, healthBarWidth + 4, 14);
 		g.setColor(new Color(237,92,80));
-		g.fillRect((Main.WIDTH / 2) - healthBarWidth / 2, Main.HEIGHT - 100, healthToWidth, 10);
+		g.fillRect((Main.WIDTH / 2) - healthBarWidth / 2, Main.HEIGHT - 100, (int) healthToWidth, 10);
 	}
 
 }
